refactor: dash-case all file names
