+++ src/modules/debug/chunk.ts
@@ -1,5 +1,5 @@
 import { BeeResponse } from '../../types'
-import { safeAxios } from '../../utils/safeAxios'
+import { safeAxios } from '../../utils/safe-axios'
 
 const endpoint = '/chunks'
