+++ src/utils/expose.ts
@@ -4,5 +4,5 @@ export * as File from './file'
 export * as Hex from './hex'
 export * as Tar from './tar'
 export * as Eth from './eth'
-export { setDefaultHeaders } from './safeAxios'
+export { setDefaultHeaders } from './safe-axios'
 export { keccak256Hash } from './hash'
